Web-Scraper for Opposition Research
A script that, when used to call a complex Custom Search API from Google, will attempt to capture all textual data from each of the top 50 resulting web-pages to any user query, parse said HTML data into legible format and output it to single document. The script is to be used for opposition research for a political campaign, to seek websites with specific text phrases beyond the capacity of a standard search engine search.