New Asset Checklist
A script utilized by all Client Field Support Imaging Technicians on newly imaged assets to ensure all encryption is in place, software is installed, and the image process is complete. It is intended to be used on groups of machines that have been freshly imaged. Field Support estimates a total of 120 hours are spent performing this process each month, and 60 hours of savings (50% man-hour reduction) is the project goal, which has been confirmed met by this solution.