Advanced Remote Troubleshooting Tool
ARTT is a collection of system administration functions designed as a proof of concept for a proposed and registered IBM intellectual property. The general concept is a single user interface for deskside technicians supporting IBM contracts to execute these administration functions on remote assets. The application itself will be cloud-based using the IBM Cloud and a database of functions, with client-side endpoints with domain admin service IDs. All IBM Technicians will one day be able to utilize an application built with this framework to execute automated functions and scripts on client enterprise workstations remotely, securely, as with full admin rights, without needing a client-side endpoint of their own. The proof of concept GUI is written in powershell calling the Windows.Forms Assembly for the time being but will eventually be rebuilt.