Upload project files
Simple calculator, built in Visual Studio using Visual Basic. NOTE: VS was used largely due to time constraints. I believe that evidence of my ability to create forms using the C# Windows.Forms Assembly has already been demonstrated in other scripts in the Master Branch, and this sample was intended to demonstrated the ability to write VB, which I feel is shown in the script subroutines.